// src/components/auth/signup/types.ts
export interface SignupFormValues {
  email: string;
  emailDomain: string;
  customDomain: string;
  password: string;
  confirmPassword: string;
  name: string;
  phone: string;
  address: string;
  addressDetail: string;
  preferjob: string[];
  preferlocate: string[];
  prefertime: string[];
  personality: string[];
  gender: string;
  birthYear: string;
  birthMonth: string;
  birthDay: string;
}